+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- partial implementation of a `File` from *dart:io* with `FileSystemFileHandle`.
+- partial implementation of a `Directory` from *dart:io* with `FileSystemDirectoryHandle`.
+- partial implementation of a `FileStat` from *dart:io* with `FileSystemFileHandle` / `FileSystemDirectoryHandle`.
+- method `FileSystemAccess.toFile` to convert a `FileSystemFileHandle` to a `File`.
+- method `FileSystemAccess.toDirectory` to convert a `FileSystemDirectoryHandle` to a `Directory`.
 
 ------------------------